+++ mss/tests/test_windows.py
@@ -52,16 +52,17 @@ def test_region_caching():
 
 
 def run_child_thread(loops):
-    # every loop will take about 1 second.
-    for i in range(loops):
+    """Every loop will take about 1 second."""
+    for _ in range(loops):
         with mss.mss() as sct:
-            mon = sct.monitors[1]
-            sct.grab(mon)
+            sct.grab(sct.monitors[1])
 
 
 def test_thread_safety():
-    """Thread safety test for issue #150."""
-    # let thread 1 finished ahead of thread2
+    """Thread safety test for issue #150.
+    The following code will throw a ScreenShotError exception if thread-safety is not guaranted.
+    """
+    # Let thread 1 finished ahead of thread 2
     thread1 = threading.Thread(target=run_child_thread, args=(30,))
     thread2 = threading.Thread(target=run_child_thread, args=(50,))
     thread1.start()